gajigaksek-backend/src/main/java/com/gjgs/gjgs/modules/utils/security/SecurityUtil.java
package com.gjgs.gjgs.modules.utils.security;

import com.gjgs.gjgs.modules.exception.member.MemberNotFoundException;
import com.gjgs.gjgs.modules.member.entity.Member;
import com.gjgs.gjgs.modules.member.enums.Authority;
import com.gjgs.gjgs.modules.member.repository.interfaces.MemberRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.stereotype.Component;

import java.util.Arrays;
import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

@Component
@Slf4j
@RequiredArgsConstructor
public class SecurityUtil {

    private final MemberRepository memberRepository;

    public Member getCurrentUserOrThrow(){
        String username = getCurrentUsername().orElseThrow(MemberNotFoundException::new);
        return memberRepository.findByUsername(username)
                .orElseThrow(MemberNotFoundException::new);
    }

    // SecurityContext 에 유저 정보가 저장되는 시점
    // Request 가 들어올 때 JwtFilter 의 doFilter 에서 저장
    public Optional<String> getCurrentUsername() {
        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();

        if (authentication == null || authentication.getPrincipal().equals("anonymousUser")) {
            return Optional.empty();
        }

        String username = null;
        if (authentication.getPrincipal() instanceof UserDetails) {
            UserDetails springSecurityUser = (UserDetails) authentication.getPrincipal();
            username = springSecurityUser.getUsername();
        } else if (authentication.getPrincipal() instanceof String) {
            username = (String) authentication.getPrincipal();
        }

        return Optional.ofNullable(username);
    }

    public Optional<Authority> getAuthority(){
        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();

        if (authentication == null || authentication.getPrincipal().equals("anonymousUser")) {
            return Optional.empty();
        }

        Authority authority = null;
        List<String> authorityList = Arrays.stream(Authority.values()).map(Enum::name).collect(Collectors.toList());

        if (authentication.getPrincipal() instanceof UserDetails) {
            UserDetails springSecurityUser = (UserDetails) authentication.getPrincipal();
            String auth = springSecurityUser.getAuthorities().toString();
            auth = auth.substring(1, auth.length() - 1);
            if (authorityList.contains(auth)){
                authority = Authority.valueOf(auth);
            }
        }
        return Optional.ofNullable(authority);
    }
}
